pyproject.toml
[project]
name = "louvores"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.2",
    "sqlmodel>=0.0.37",
]
